+++ src/components/Property/RaiseWorkOrder/RaiseWorkOrderFollowOn/FollowOnLookup.tsx
@@ -1,94 +1,30 @@
 import { useEffect, useState } from 'react'
-import Radios from '../../Form/Radios'
-import { WorkOrder } from '@/root/src/models/workOrder'
+import ErrorMessage from '../../../Errors/ErrorMessage'
+import { DataList } from '../../../Form'
+import SpinnerWithLabel from '../../../SpinnerWithLabel'
 import { frontEndApiRequest } from '@/root/src/utils/frontEndApiClient/requests'
-import { DataList } from '../../Form'
-import SpinnerWithLabel from '../../SpinnerWithLabel'
-import ErrorMessage from '../../Errors/ErrorMessage'
+import { WorkOrder } from '@/root/src/models/workOrder'
 
 interface Props {
-  register: any
-  watch: any
-  propertyReference: string
-  errors: { [key: string]: { message: string } }
-}
-
-const RADIO_OPTIONS = [
-  { text: 'Yes', value: true },
-  { text: 'No', value: false },
-]
-
-const RaiseWorkOrderFollowOn = (props: Props) => {
-  const { register, errors, watch, propertyReference } = props
-
-  const [showFollowOnLookup, setShowFollowOnLookup] = useState<boolean>(false)
-
-  const isFollowOnWatchedValue = watch('isFollowOn')
-
-  useEffect(() => {
-    setShowFollowOnLookup(() => isFollowOnWatchedValue == 'true')
-  }, [isFollowOnWatchedValue])
-
-  return (
-    <div
-      style={{
-        background: '#f9f9f9',
-        padding: '15px',
-      }}
-    >
-      <Radios
-        name="isFollowOn"
-        label="Is this for follow on works?"
-        labelSize={'s'}
-        options={RADIO_OPTIONS.map((x) => ({
-          ...x,
-          defaultChecked: false,
-          children:
-            x.value === true ? (
-              <FollowOnLookup
-                visible={showFollowOnLookup}
-                propertyReference={propertyReference}
-                register={register}
-                errors={errors}
-              />
-            ) : null,
-        }))}
-        error={errors && errors.isFollowOn}
-        register={register({
-          required: 'Please select an option',
-        })}
-      />
-
-      <div>
-        <a
-          className="lbh-link"
-          href={`/search`}
-          target="_blank"
-          rel="noreferrer"
-        >
-          Look up work order number
-        </a>
-      </div>
-    </div>
-  )
-}
-
-const FollowOnLookup = ({
-  visible,
-  propertyReference,
-  register,
-  errors,
-}: {
   visible: boolean
   propertyReference: string
   register: any
   errors: { [key: string]: { message: string } }
-}) => {
+}
+
+const FollowOnLookup = (props: Props) => {
+  const { visible, propertyReference, register, errors } = props
+
   const [workOrders, setWorkOrders] = useState<WorkOrder[]>([])
   const [isLoading, setIsLoading] = useState<boolean>(false)
   const [error, setError] = useState<string | null>(null)
 
   const fetchRelatedWorkOrders = async () => {
+    if (workOrders.length > 0) {
+      // we dont need to fetch new data
+      return
+    }
+
     setError(null)
     setIsLoading(true)
 
@@ -139,6 +75,7 @@ const FollowOnLookup = ({
 
   return (
     <div>
+      {/* @ts-expect-error Props validation */}
       <DataList
         name="parentWorkOrder"
         label="Please select the original work order this relates to"
@@ -156,4 +93,4 @@ const FollowOnLookup = ({
   )
 }
 
-export default RaiseWorkOrderFollowOn
+export default FollowOnLookup
